requirements.txt
streamlit
python-dotenv
PyPDF2
streamlit-extras
langchain
langchain-openai
faiss-cpu
